+++ deepxube/base/domain.py
@@ -1,13 +1,18 @@
 from abc import ABC, abstractmethod
 from typing import List, Tuple, Optional, Set, TypeVar, Generic, Dict, Any
+
+import torch
+from torch import nn
 import numpy as np
 from clingo.solving import Model as ModelCl
 
 from deepxube.logic.logic_objects import Atom, Model
 from deepxube.utils import misc_utils
-from deepxube.nnet.nnet_utils import NNetPar, NNetCallable
+from deepxube.nnet.nnet_utils import NNetPar, NNetCallable, load_nnet
 from deepxube.utils.timing_utils import Times
+
 from matplotlib.figure import Figure
+import os
 import random
 import time
 from numpy.typing import NDArray
@@ -71,7 +76,8 @@ G = TypeVar('G', bound=Goal)
 # TODO method for downloading data?
 class Domain(ABC, Generic[S, A, G]):
     def __init__(self, *args: Any, **kwargs: Any) -> None:
-        self.nnet_pars: List[Tuple[str, str, NNetPar]] = []
+        self.nnet_par_dict: Dict[str, Tuple[str, NNetPar]] = dict()
+        self.nnet_fn_dict: Dict[str, NNetCallable] = dict()
 
     @abstractmethod
     def sample_problem_instances(self, num_steps_l: List[int], times: Optional[Times] = None) -> Tuple[List[S], List[G]]:
@@ -155,11 +161,37 @@ class Domain(ABC, Generic[S, A, G]):
 
         return states_walk, actions_l, path_costs
 
-    def get_nnet_pars(self) -> List[Tuple[str, str, NNetPar]]:
-        return self.nnet_pars
+    def get_nnet_par_dict(self) -> Dict[str, Tuple[str, NNetPar]]:
+        return self.nnet_par_dict.copy()
 
     def set_nnet_fns(self, nnet_fn_dict: Dict[str, NNetCallable]) -> None:
-        pass
+        for nnet_name, nnet_fn in nnet_fn_dict.items():
+            if nnet_name in self.nnet_par_dict.keys():
+                self.nnet_fn_dict[nnet_name] = nnet_fn
+
+    def get_nnet_fn(self, nnet_fn_name: str) -> NNetCallable:
+        nnet_fn: Optional[NNetCallable] = self.nnet_fn_dict.get(nnet_fn_name)
+        if nnet_fn is None:
+            device: torch.device = torch.device("cpu")
+            if ('CUDA_VISIBLE_DEVICES' in os.environ) and torch.cuda.is_available():
+                device = torch.device("cuda:%i" % 0)
+
+            assert nnet_fn_name in self.nnet_par_dict.keys(), f"nnet_fn_name {nnet_fn_name} has not been added to nnet_par_dict"
+            nnet_file, nnet_par = self.nnet_par_dict[nnet_fn_name]
+            nnet: nn.Module = nnet_par.get_nnet()
+            nnet = load_nnet(nnet_file, nnet, device=device)
+            nnet.to(device)
+
+            nnet_fn = nnet_par.get_nnet_fn(nnet, None, device, None)
+
+        return nnet_fn
+
+    def _add_nnet_par(self, nnet_name: str, nnet_file: str, nnet_par: NNetPar) -> None:
+        self.nnet_par_dict[nnet_name] = (nnet_file, nnet_par)
+
+    def __getstate__(self):
+        self.nnet_fn_dict = dict()
+        return self.__dict__
 
 
 # Visualization mixins