+++ CHANGES.md
@@ -19,6 +19,8 @@
 * get_path returns list of transition costs
 * Added rollout pathfinding alg, special case of beam search that does not check is_solved when searching
 * PolicyNNet has _forward_train and _forward_eval abstract methods to unify forward method
+* Add checkpointing during training
+* Add more functionality to add neural network functions to domain and use in parallel
 
 ## 0.2.1
 * Consolidate search: Beam search -> special cases: greedy_policy, graph search -> special cases: batch weighted A* search, batch weighted Q* search